Update bot and add documentation
