+++ docs/user/bots.rst
@@ -3123,6 +3123,27 @@ Use this command to create/update the database and reload the bot:
 
 .. _intelmq.bots.experts.url2fqdn.expert:
 
+Truncate By Delimiter
+^^^^^^^^^
+
+**Information**
+
+* `name:` `intelmq.bots.experts.truncate_by_delimiter.expert`
+* `lookup:` no
+* `public:` yes
+* `cache (redis db):` none
+* `description:` Cut string if length is bigger than max
+
+**Configuration Parameters**
+
+* `delimiter`: example . or ;
+* `max_length`: max string length
+* `field`: string field
+
+Example: Cut through a long domain with a dot.
+
+.. _intelmq.bots.experts.truncate_by_delimiter.expert:
+
 Url2FQDN
 ^^^^^^^^
